requirements.txt
pynetworktables==2019.0.2
tornado==6.0.3
zmq=0.0.0
